requirements.txt
Django>=5.1.7
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.7
PyYAML>=6.0.1  # drf-yasg에서 내부적으로 필요
